pyproject.toml
[tool.poetry]
name = "matching-cv-and-vacancy"
version = "0.1.0"
description = ""
authors = ["Anton Shiryaev <med.phisiker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
numpy = "^1.26.2"
ipykernel = "^6.27.1"
lxml = "^4.9.3"
sentence-transformers = "^2.2.2"
streamlit = "^1.29.0"
faiss-gpu = {version = "^1.7.2", platform = "linux"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
